add docker and configurations project

+++ src/main/kotlin/com/dsmanioto/searchopportunity/searchopportunityapi/model/Opportunity.kt
@@ -1,3 +1,4 @@
 package com.dsmanioto.searchopportunity.searchopportunityapi.model
 
-data class Opportunity(val id: Int, val description: String)
+data class Opportunity(val id: Long,
+                       val description: String)
